Fix #55
